last chekc to do then done(idk what could break this)

+++ srcs/firstchild.c
@@ -6,7 +6,7 @@
 /*   By: afrigger <afrigger@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/28 13:49:54 by afrigger          #+#    #+#             */
-/*   Updated: 2023/02/21 12:36:48 by afrigger         ###   ########.fr       */
+/*   Updated: 2023/02/22 11:41:54 by afrigger         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,6 +37,8 @@ void	firstchild(t_pipe *pipex, char **envp, char**args, int fd[2])
 	set_args(pipex, args[2]);
 	testpath(pipex);
 	filecheck(args[1]);
+	if (pipex->pathok == NULL)
+		cmdnotfound(pipex);
 	error = execve(pipex->pathok, pipex->cmd1_args, envp);
 	if (error == -1)
 		errcheck(pipex);
@@ -46,18 +48,32 @@ void	secondchild(t_pipe *pipex, char **envp, char**args, int fd[2])
 {
 	int	output;
 	int	error;
-
-	unlink(args[4]);
-	output = open(args[4], O_WRONLY | O_CREAT, 0777);
-	dup2(output, STDOUT_FILENO);
+	if(args[4])
+	{
+		unlink(args[4]);
+		output = open(args[4], O_WRONLY | O_CREAT, 0777);
+		dup2(output, STDOUT_FILENO);	
+		close(output);	
+	}
 	dup2(fd[0], STDIN_FILENO);
 	close(fd[1]);
 	close(fd[0]);
-	close(output);
 	splitpath(envp, pipex);
 	set_args(pipex, args[3]);
 	testpath(pipex);
+	if (pipex->pathok == NULL)
+		cmdnotfound(pipex);
 	error = execve(pipex->pathok, pipex->cmd1_args, envp);
 	if (error == -1)
 		errcheck(pipex);
+}
+
+
+void	cmdnotfound(t_pipe *pipex)
+{
+	ft_putstr_fd("pipex: ", 2);
+	ft_putstr_fd("command not found: ", 2);
+	ft_putstr_fd(pipex->cmd1, 2);
+	ft_putstr_fd("\n", 2);
+	exit(1);
 }
